Refactor pipe connection handling and separate capability registrations
