+++ fabric/src/main/java/com/logistics/fabric/capability/PipeConnectionAccess.java
@@ -1,7 +1,6 @@
-package com.logistics.core.lib.block.lookup;
+package com.logistics.fabric.capability;
 
 import com.logistics.core.lib.block.capability.PipeConnection;
-import com.logistics.core.lib.pipe.PipeConnectionRegistry;
 
 /**
  * Bridges internal {@link PipeConnection} interface to the PipeConnectionRegistry lookup API.
